Avoid ambiguous overloads in BitVector
`long` is a 32-bit integer on Windows. CVC4's BitVector class had a
constructor for `unsigned int` and `unsigned long`, which lead to issues
with the new CVC4 C++ API because the two constructors were ambiguous
overloads. This commit changes the constructors to use `uint32_t` and
`uint64_t`, which are plattform independent and more explicit (mirroring
the API).

+++ src/theory/fp/fp_converter.cpp
@@ -275,7 +275,7 @@ symbolicProposition symbolicRoundingMode::valid(void) const
 {
   NodeManager *nm = NodeManager::currentNM();
   Node zero(nm->mkConst(
-      BitVector(SYMFPU_NUMBER_OF_ROUNDING_MODES, (long unsigned int)0)));
+      BitVector(SYMFPU_NUMBER_OF_ROUNDING_MODES, static_cast<uint32_t>(0))));
 
   // Is there a better encoding of this?
   return symbolicProposition(nm->mkNode(
@@ -287,7 +287,7 @@ symbolicProposition symbolicRoundingMode::valid(void) const
                                        *this,
                                        nm->mkConst(BitVector(
                                            SYMFPU_NUMBER_OF_ROUNDING_MODES,
-                                           (long unsigned int)1)))),
+                                           static_cast<uint32_t>(1))))),
                  zero),
       nm->mkNode(kind::BITVECTOR_NOT,
                  nm->mkNode(kind::BITVECTOR_COMP, *this, zero))));
